Simplify grails-transactional-events v8: listeners are plain services
The v8 guide registered its four Spring event listeners one-by-one in
resources.groovy (behind a long explanatory comment) because the classes
were named *Listener and so were not Grails artefacts. Renaming them to
*Service lets Grails auto-register them by convention - the simpler,
idiomatic approach:

- CustomerLifetimeValueListener -> CustomerLifetimeValueService
- AuditListener               -> AuditService
- NotificationListener        -> NotificationService
- WorkOrderAssignmentListener -> WorkOrderPlanningService

resources.groovy now declares only the TransactionalEventListenerFactory
bean - the one piece Grails genuinely does not provide, because it drives
transactions through GORM's @Transactional rather than Spring's
@EnableTransactionManagement.

The narrative is updated throughout: the registration chapter also explains
the src/main/groovy + @ComponentScan alternative, and the AFTER_COMMIT
guidance consistently uses withNewTransaction (Grails' @Transactional AST
transform would otherwise hide a listener method from Spring's scanner).
