+++ guides/grails-transactional-events/v8/guide/integrationTest.adoc
@@ -11,7 +11,7 @@ A few details that catch people the first time:
 * *Why the spec is not annotated `@Rollback`.* The default integration-test behaviour in Grails wraps each test method in a transaction that's rolled back after the method returns. If you let that wrapper roll back the *outer* transaction, the AFTER_COMMIT listeners inside `placeOrder(...)` never fire - and the test cannot distinguish "the listener didn't run because of rollback semantics (good)" from "the listener didn't run because the test wrapper rolled back (irrelevant)". The spec therefore opens its own committed transaction with `Order.withNewTransaction { ... }` and cleans up explicitly.
 * *Why `setRollbackOnly()` instead of throwing.* Throwing an exception from inside `withNewTransaction { ... }` would propagate out of the test method and fail the spec. `status.setRollbackOnly()` tells Spring to roll back when the block exits normally, which is exactly the "user-initiated rollback" scenario the test wants to model.
 * *Why the `cleanup:` block uses `withNewTransaction` again.* The test ran outside the integration-test transaction wrapper, so there is no automatic rollback to lean on. The cleanup must explicitly delete the rows the test created, in their own transaction, in the right order (`AuditLog` rows first because they reference `orderId`, then `Order`, then `Customer`).
-* *Why the async `NotificationListener` is not asserted on.* It logs only - there is nothing in the database to assert against. Testing that an `@Async` listener actually ran requires either an `awaitility`-style poll or replacing the executor with a synchronous one in the test profile. The two non-async listeners are sufficient to demonstrate the AFTER_COMMIT contract.
+* *Why the async `NotificationService` is not asserted on.* It logs only - there is nothing in the database to assert against. Testing that an `@Async` listener actually ran requires either an `awaitility`-style poll or replacing the executor with a synchronous one in the test profile. The two non-async listeners are sufficient to demonstrate the AFTER_COMMIT contract.
 
 Run the spec:
