+++ guides/grails-transactional-events/v8/snippets/grails-app/services/example/AuditService.groovy
@@ -7,16 +7,17 @@ import org.springframework.transaction.event.TransactionalEventListener
 /**
  * Persists an AuditLog row for every committed order.
  *
- * A separate listener class - not a method on OrderService - keeps the
- * "place an order" concern decoupled from the "record audit history"
- * concern. Adding another committed-only side-effect later means a new
- * class, not an edit to OrderService.
+ * A separate service - not a method on OrderService - keeps the "place an
+ * order" concern decoupled from the "record audit history" concern. Adding
+ * another committed-only side-effect later means a new service, not an edit to
+ * OrderService.
  *
- * Like CustomerLifetimeValueListener, it is registered in resources.groovy
- * and does its GORM write inside withNewTransaction { } because the
- * AFTER_COMMIT callback fires after the publisher's transaction has committed.
+ * Like CustomerLifetimeValueService it is auto-registered by Grails (its name
+ * ends in 'Service') and does its GORM write inside withNewTransaction { }
+ * because the AFTER_COMMIT callback fires after the publisher's transaction
+ * has committed.
  */
-class AuditListener {
+class AuditService {
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     void onOrderPlaced(OrderPlacedEvent event) {
