+++ guides/grails-transactional-events/v8/snippets/grails-app/services/example/NotificationService.groovy
@@ -10,22 +10,22 @@ import org.springframework.transaction.event.TransactionalEventListener
 /**
  * Off-thread side-effect: stand-in for sending a confirmation email.
  *
- * @Async dispatches the listener body to Spring's task executor instead
- * of running it on the caller's thread. The transaction-phase binding
- * is still synchronous - Spring waits for AFTER_COMMIT, then hands the
- * event to the executor. The listener body therefore has NO inherited
- * transaction; if it needs the database it must open its own via
- * @Transactional(propagation = REQUIRES_NEW).
+ * An ordinary Grails service, auto-registered by name. @Async dispatches the
+ * listener body to Spring's task executor instead of running it on the
+ * caller's thread. The transaction-phase binding is still synchronous - Spring
+ * waits for AFTER_COMMIT, then hands the event to the executor. The body
+ * therefore has NO inherited transaction; if it needs the database it must
+ * open its own via withNewTransaction { }.
  *
  * Exceptions thrown from an @Async listener are NOT propagated to the
  * publisher - see Spring's AsyncUncaughtExceptionHandler.
  *
- * @EnableAsync must be present on Application.groovy for @Async to take
- * effect; without it the annotation is silently a no-op.
+ * @EnableAsync must be present on Application for @Async to take effect;
+ * without it the annotation is silently a no-op.
  */
-class NotificationListener {
+class NotificationService {
 
-    private static final Logger log = LoggerFactory.getLogger(NotificationListener)
+    private static final Logger log = LoggerFactory.getLogger(NotificationService)
 
     @Async
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
